POSDAO stabilization + Randao duplicate transactions fix (#2095)
* This changes are temporary - needed to investigate issue in RANDAO that produced multiple transactions with same nonce for same phase in block.

* Move finalization info log Debug->Trace

* review fixes

* Fix NullReference on failed to sign block

* Fix nonces in block when adding multiple engine transactions

* Remove some unneded code, refactor and improve some loging

* Don't let generated transactions into transaction pool

* v2. Don't switch to full sync when got wrong TotalDifficulty but same HeaderHash - filter peers

* Ignore duplicate reports for same block

* Allow generated transactions into TxPool (for report transactions)

* fix receipt migration message

* increase blockTooEarlyWarningMillisecondThreshold to 1s

* make reporting unique per multiple classes and threads

* add logging of included transactions to blocks

* Benign report transactions not PersistentBroadcast

* disable filtering on duplicate reports for testing

* Add stack trace to report for debugging

* Fix AlreadyReported

* change to more general exception handling during report

* Improve AlreadyReported check

* change logging

* added logging

* adding logging

* small fix on ProcessingQueueEmpty

* LOGS

* Logs + better reporting discovery

* Add step-sealer validation on block discovery when head is processed

* temp logs

* temp logs

* Added PeerClientType.OpenEthereum,

Added AuRaMultiSyncModeSelector ignoring outliers from Parity/OpenEthereum

* Ignore also parents of our head in MultiSyncModeSelector peer selection

* even more logging

* Fix multithreading issue in BaseLoopBlockProducer

* Fix total difficulty from parity

* Adding metrics for processing blocks

* Adding AuRa metrics

* cleanup

* review fixes

+++ src/Nethermind/Nethermind.Blockchain/Processing/ProcessingStats.cs
@@ -61,7 +61,7 @@ namespace Nethermind.Blockchain.Processing
 
             Metrics.Mgas += block.GasUsed / 1_000_000m;
             Metrics.Transactions += block.Transactions.Length;
-            Metrics.Blocks = (long)block.Number;
+            Metrics.Blocks = block.Number;
             Metrics.RecoveryQueueSize = recoveryQueueSize;
             Metrics.ProcessingQueueSize = blockQueueSize;
